shadowfix/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pillow
python-multipart
opencv-python-headless
pydantic-settings
python-jose[cryptography]
bcrypt
slowapi
gunicorn
